src/requirements.txt
# I have Version locked as per current requirements can be updated
azure-cosmos==4.6.0
azure-storage-blob==12.19.1
